# docs/architecture/contracts/rest/openapi.yaml
openapi: 3.1.0
info:
  title: Mintkey Admin REST API
  version: "0.1.0-experimental"
  summary: Operator-facing control-plane HTTP surface for Mintkey.
  description: |
    Mintkey is a credential broker for AI agents. This API is consumed by:
      - The Admin Console (AdminJS) for sensitive operations.
      - Operator-owned tooling (CI, scripts) using API tokens.
      - Internal subscribers calling `/v1/changes` for reconciliation
        (per ADR-0010 change-channel model).

    Stability tier: **experimental**. Breaking changes will bump the major
    `info.version`; surface stability is per-resource and called out in each
    operation `description` where it differs from the document-level tier.

    ## ID conventions

    All identifiers are **ULIDs** (Crockford base32, 26 characters) prefixed
    by their resource type. Tenants are addressed at the wire by their prefixed ULID `tenant_…` on
    every entity for RLS purposes (ADR-0008).

    | Prefix          | Resource                          |
    | --------------- | --------------------------------- |
    | `tenant_…`      | Tenant                            |
    | `operator_…`    | Operator                          |
    | `agent_…`       | Agent                             |
    | `svc_…`         | Service                           |
    | `cred_…`        | Credential (key-version envelope) |
    | `perm_…`        | PermissionGrant                   |
    | `audit_…`       | AuditEvent                        |
...

...
servers:
...
security:
  - OperatorSession: []
  - OperatorBearer: []
paths:
...
  /v1/tenants/{tenant_id}/services/{service_id}/credentials:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/ServiceId"
    get:
      tags: [Credentials]
      operationId: listCredentialVersions
      summary: List credential versions for a service.
      description: |
        Returns metadata only; the plaintext value field is `null` in every
        item. To reveal a credential, the operator must rotate
        (`POST /credentials`); the plaintext is returned exactly once on the
        rotation response.
      parameters:
        - $ref: "#/components/parameters/AfterCursor"
        - $ref: "#/components/parameters/Limit"
        - in: query
          name: q
          required: false
          schema:
            type: string
            maxLength: 200
          description: >
            Case-insensitive substring search on auth_scheme.
            Note: credentials have no human-readable name field; q is limited
            to auth_scheme matching.
      responses:
        "200":
          description: A page of credential versions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialPage"
              example:
                data:
                  - id: "cred_01HX5J9F8V8H8V0CG3F2Y5J6C3"
                    tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                    service_id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S1"
                    key_version: 3
                    auth_scheme: "bearer_token"
                    is_current: true
                    status: "active"
                    value: null
                    created_at: "2026-05-08T16:22:00Z"
                    updated_at: "2026-05-08T16:22:00Z"
                  - id: "cred_01HX5J9F8V8H8V0CG3F2Y5J6C2"
                    tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                    service_id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S1"
                    key_version: 2
                    auth_scheme: "bearer_token"
                    is_current: false
                    status: "active"
                    value: null
                    created_at: "2026-05-01T08:00:00Z"
                    updated_at: "2026-05-01T08:00:00Z"
                next_cursor: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Credentials]
      operationId: registerOrRotateCredential
      summary: Register or rotate a credential.
      description: |
        Always creates a new `key_version` (monotonic per service). The
        plaintext value is returned **exactly once**, in this response.
        Subsequent reads return `value: null`. The previous key_version
        is retained until soft-deleted (see `DELETE`).

        The credential is stored encrypted via the Vault Adapter
        (ADR-0003). A `credential.rotated` audit event is emitted; a
        `credential.rotated` change event is published on
        `mintkey:credential` (tenant_id is in payload) (ADR-0010).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCredentialRequest"
            examples:
              api-key-header:
                value:
                  auth_scheme: "api_key_header"
                  value: "<redacted-example-api-key>"
                  header_name: "X-API-Key"
              bearer-token:
                value:
                  auth_scheme: "bearer_token"
                  value: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              basic-auth:
                value:
                  auth_scheme: "basic_auth"
                  value: "alice:s3cret"
              oauth2-cc:
                value:
                  auth_scheme: "oauth2_client_credentials"
                  client_id: "abc"
                  client_secret: "shh"
                  token_url: "https://oauth.example/token"
                  scopes: ["api.read"]
      responses:
        "201":
          description: Credential created. Plaintext returned exactly once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
              example:
                id: "cred_01HX5J9F8V8H8V0CG3F2Y5J6C4"
                tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                service_id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S1"
                key_version: 4
                auth_scheme: "bearer_token"
                is_current: true
                status: "active"
                value: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                created_at: "2026-05-10T14:00:00Z"
                updated_at: "2026-05-10T14:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/tenants/{tenant_id}/services/{service_id}/credentials/{key_version}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/ServiceId"
      - in: path
        name: key_version
        required: true
        schema:
          type: integer
          minimum: 1
        description: Monotonic credential version (1-indexed).
    delete:
      tags: [Credentials]
      operationId: revokeCredentialVersion
      summary: Soft-delete a non-current credential version.
      description: |
        Refuses to delete `is_current = true`. Use rotation
        (`POST /credentials`) to advance the current version; the prior
        version may then be soft-deleted via this endpoint.

        Emits a `credential.revoked` audit event and a `credential.revoked`
        change event.
      responses:
        "204":
          description: Soft-deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            Credential is the current version and cannot be deleted.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: "about:blank"
                title: "Conflict"
                status: 409
                detail: "Cannot delete current credential version 4. Rotate first."
                instance: "/v1/tenants/tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9/services/svc_01HX5J9F8V8H8V0CG3F2Y5J6S1/credentials/4"
                "mintkey:code": "credential_in_use"
                "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"

  ###########################################################################
  # Agents
  ###########################################################################
...
components:
...
  securitySchemes:
    OperatorSession:
      type: apiKey
      in: cookie
      name: mintkey_session
      description: |
        HttpOnly Secure SameSite=Strict session cookie issued by
        `/v1/auth/internal-login` or `/v1/auth/callback`.
    OperatorBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Operator-owned API token (PAT) for machine-to-machine calls. Issued
        from the Admin Console; not exposed in this OpenAPI document. Sent
        as `Authorization: Bearer …`.
    AgentApiKey:
      type: http
      scheme: bearer
      bearerFormat: mintkey-agent-api-key
      description: |
        Long-lived Agent API Key (`mk_live_…`). Used by the few agent-facing
        REST endpoints; most agent traffic flows through MCP instead.

    AdminUiSignedRequest:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Ed25519 JWT signed by the AdminJS service. Required on every endpoint
        AdminJS calls on behalf of an operator. Carries claims:
          iss: "mintkey/admin-ui"
          sub: <operator_id>
          tnt: <tenant_id>
          aud: "mintkey/admin-api"
          iat, exp (60s TTL)
          jti (UUID; replay-checked against admin_request_jti per ADR-0016.1)
        FastAPI verifies the signature against AdminJS's public key fetched from
        the Vault Adapter under credential type 'admin_ui_signing_key'.
        Per ADR-0014.6 + ADR-0017.1.
    ServiceIdentity:
...
  parameters:
...
    TenantId:
      in: path
      name: tenant_id
      required: true
      schema:
        type: string
        pattern: "^tenant_[0-9A-HJKMNP-TV-Z]{26}$"
      description: ULID-prefixed tenant identifier.
    ServiceId:
      in: path
      name: service_id
      required: true
      schema:
        type: string
        pattern: "^svc_[0-9A-HJKMNP-TV-Z]{26}$"
      description: ULID-prefixed service identifier.
...
    AfterCursor:
      in: query
      name: after
      required: false
      schema:
        type: string
      description: |
        Cursor — the `id` of the last item from the previous page. Exclusive.
    Limit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Max items to return (1..200, default 50).

  ###########################################################################
  # Reusable error responses
  ###########################################################################
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Bad Request"
            status: 400
            detail: "`base_url` must be an absolute https URL."
            instance: "/v1/tenants/tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9/services"
            "mintkey:code": "validation_failed"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
            errors:
              - field: "base_url"
                message: "must be an absolute https URL"
    Unauthorized:
      description: No or invalid credentials.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Unauthorized"
            status: 401
            detail: "Session cookie missing or invalid."
            instance: "/v1/auth/whoami"
            "mintkey:code": "unauthenticated"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
    Forbidden:
      description: Authenticated but not allowed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Forbidden"
            status: 403
            detail: "Operator role 'Auditor' cannot create services."
            instance: "/v1/tenants/tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9/services"
            "mintkey:code": "permission_denied"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
    NotFound:
      description: Resource does not exist or is not visible to the caller.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Not Found"
            status: 404
            detail: "No tenant with id 'tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9'."
            instance: "/v1/tenants/tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
            "mintkey:code": "tenant_not_found"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
    Conflict:
      description: State conflict.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Conflict"
            status: 409
            detail: "A service with name 'crm' already exists in this tenant."
            instance: "/v1/tenants/tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9/services"
            "mintkey:code": "duplicate_resource"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
...
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 7807 problem document with two Mintkey-specific extensions:
        `mintkey:code` (machine-readable error code) and
        `mintkey:trace_id` (W3C `trace-id` for correlation).
      properties:
        type:
          type: string
          format: uri-reference
          default: "about:blank"
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        "mintkey:code":
          type: string
          description: Machine-readable error code (snake_case).
          examples:
            - "tenant_not_found"
            - "permission_denied"
            - "credential_in_use"
            - "duplicate_resource"
            - "validation_failed"
            - "unauthenticated"
            - "rate_limited"
            - "change_channel_unavailable"
            - "vault_unavailable"
            - "oidc_unavailable"
        "mintkey:trace_id":
          type: string
          pattern: "^[0-9a-f]{32}$"
          description: W3C `trace-id` (32 hex chars).
        errors:
          type: array
          description: Optional list of per-field validation errors.
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
            required: [field, message]
      required: [title, status]

    # ----- Common identifiers -----
...
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 UTC timestamp.

    # ----- Tenant -----
...
    AuthScheme:
      type: string
      enum:
        - api_key_header
        - api_key_query
        - bearer_token
        - basic_auth
        - oauth2_client_credentials
        - oidc_client_secret
      description: |
        How Mintkey injects the credential into the outbound request.
        - `api_key_header`: header injection (e.g. `X-API-Key: …`).
        - `api_key_query`: query-parameter injection (e.g. `?api_key=…`).
        - `bearer_token`: `Authorization: Bearer …`.
        - `basic_auth`: `Authorization: Basic base64(user:pass)`.
        - `oauth2_client_credentials`: token-exchange flow.
        - `oidc_client_secret`: used for the operator IdP integration only.
...
    Credential:
      type: object
      properties:
        id:
          type: string
          pattern: "^cred_[0-9A-HJKMNP-TV-Z]{26}$"
        tenant_id:
          type: string
          pattern: "^tenant_[0-9A-HJKMNP-TV-Z]{26}$"
        service_id:
          type: string
          pattern: "^svc_[0-9A-HJKMNP-TV-Z]{26}$"
        key_version:
          type: integer
          minimum: 1
        auth_scheme:
          $ref: "#/components/schemas/AuthScheme"
        is_current:
          type: boolean
        status:
          type: string
          enum: [active, revoked]
        value:
          type: [string, "null"]
          x-mintkey-sensitive: true
          description: |
            Plaintext credential value. Returned **only** on the immediate
            POST response (rotation/registration). All subsequent reads
            return `null`.
        expires_at:
          type: [string, "null"]
          format: date-time
          description: |
            Optional upstream expiry (e.g. for OAuth2 client_credentials
            tokens cached in the Vault Adapter).
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
      required:
        - id
        - tenant_id
        - service_id
        - key_version
        - auth_scheme
        - is_current
        - status
        - created_at
        - updated_at
    RegisterCredentialRequest:
      description: |
        Discriminated by `auth_scheme`. Plaintext fields are write-only and
        marked `x-mintkey-sensitive: true`.
      oneOf:
        - $ref: "#/components/schemas/CredApiKeyHeader"
        - $ref: "#/components/schemas/CredApiKeyQuery"
        - $ref: "#/components/schemas/CredBearerToken"
        - $ref: "#/components/schemas/CredBasicAuth"
        - $ref: "#/components/schemas/CredOAuth2ClientCredentials"
        - $ref: "#/components/schemas/CredOidcClientSecret"
      discriminator:
        propertyName: auth_scheme
        mapping:
          api_key_header: "#/components/schemas/CredApiKeyHeader"
          api_key_query: "#/components/schemas/CredApiKeyQuery"
          bearer_token: "#/components/schemas/CredBearerToken"
          basic_auth: "#/components/schemas/CredBasicAuth"
          oauth2_client_credentials: "#/components/schemas/CredOAuth2ClientCredentials"
          oidc_client_secret: "#/components/schemas/CredOidcClientSecret"
    CredApiKeyHeader:
      type: object
      properties:
        auth_scheme:
          type: string
          enum: [api_key_header]
        value:
          type: string
          x-mintkey-sensitive: true
        header_name:
          type: string
          default: "X-API-Key"
      required: [auth_scheme, value]
    CredApiKeyQuery:
      type: object
      properties:
        auth_scheme:
          type: string
          enum: [api_key_query]
        value:
          type: string
          x-mintkey-sensitive: true
        query_param:
          type: string
          default: "api_key"
      required: [auth_scheme, value]
    CredBearerToken:
      type: object
      properties:
        auth_scheme:
          type: string
          enum: [bearer_token]
        value:
          type: string
          x-mintkey-sensitive: true
      required: [auth_scheme, value]
    CredBasicAuth:
      type: object
      properties:
        auth_scheme:
          type: string
          enum: [basic_auth]
        value:
          type: string
          description: "Format: `username:password`."
          x-mintkey-sensitive: true
      required: [auth_scheme, value]
    CredOAuth2ClientCredentials:
      type: object
      properties:
        auth_scheme:
          type: string
          enum: [oauth2_client_credentials]
        client_id:
          type: string
        client_secret:
          type: string
          x-mintkey-sensitive: true
        token_url:
          type: string
          format: uri
        scopes:
          type: array
          items:
            type: string
        audience:
          type: [string, "null"]
      required: [auth_scheme, client_id, client_secret, token_url]
    CredOidcClientSecret:
      type: object
      description: |
        OIDC client_secret used by the *broker itself* to integrate with the
        operator IdP. Only PlatformAdmin should register one.
      properties:
        auth_scheme:
          type: string
          enum: [oidc_client_secret]
        client_id:
          type: string
        client_secret:
          type: string
          x-mintkey-sensitive: true
        issuer:
          type: string
          format: uri
      required: [auth_scheme, client_id, client_secret, issuer]
    CredentialPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Credential"
        next_cursor:
          type: [string, "null"]
      required: [data, next_cursor]

    # ----- Agent -----
...